Fixes missing LyricCameraHAL sources build errors
When rebuilding the vendor image, the 'google/services/LyricCameraHAL'
directory will not be present in the AOSP 'vendor' folder, triggering
a build error when trying to copy the
'vendor.android.hardware.camera.preview-dis.xml' file.

To avoid this, the wildcard function will be used to check for the
folder's existence, adding the XML file mentioned above to the
'PRODUCT_COPY_FILES' variable only if the
'vendor/google/services/LyricCameraHAL/src' directory exists.